+++ CHANGELOG.md
@@ -26,6 +26,7 @@ full message for the in-app changelog viewer.
 
 ### Documentation
 
+- docs: add Hong Kong Cantonese sections, sixth wave - [`65ad7f4be3`](https://github.com/Ding-Ding-Projects/worldlens/commit/65ad7f4be34b36273927c7bbf212c010e16a8a7a)
 - docs: add Hong Kong Cantonese sections, fifth wave - [`5b5f89a5c1`](https://github.com/Ding-Ding-Projects/worldlens/commit/5b5f89a5c12d5b13676fa896a43f8f25184039ff)
 - docs: add Hong Kong Cantonese sections, fourth wave - [`45f0ee2515`](https://github.com/Ding-Ding-Projects/worldlens/commit/45f0ee251599345fd5ccf4d36bccdd3adcb45f34)
 - test(identity): pin the Pages legacy marker to its bilingual site count - [`00fc212a40`](https://github.com/Ding-Ding-Projects/worldlens/commit/00fc212a40606c875718cb95b93a4b459900938b)